requirements.txt
streamlit
langchain
langchain-huggingface
langchain-community
langchain-groq
faiss-cpu
transformers
python-dotenv